fix: Return Numeric type for aggregate functions (COUNT, SUM, AVG) (#1175)
* fix: Return Numeric type for aggregate functions (COUNT, SUM, AVG)

Changes aggregate functions to return SqlValue::Numeric instead of
SqlValue::Integer for consistency with SQLLogicTest expectations.

Changes:
- Updated COUNT to return Numeric in grouping.rs
- Updated COUNT, SUM, AVG to return Numeric in window aggregates
- Fixed COUNT(*) fast paths to return Numeric
- Updated test expectations to match new return types

This ensures aggregate results format with .000 decimal suffix
as expected by SQLLogicTest, fixing formatting inconsistencies.

Fixes #1171



* fix: Update test assertions for Numeric aggregate return types

Fixed 31 out of 34 test assertions to expect Numeric/Float types
instead of Integer for COUNT, SUM, and AVG aggregate results.

Changes:
- Updated join_aggregation.rs: COUNT assertions to Numeric
- Updated select_window_aggregate.rs: window aggregate assertions
- Updated aggregate_caching.rs: COUNT assertions to Float
- Updated count_star_fast_path.rs: fast path COUNT assertions
- Updated aggregate_without_from.rs: aggregate expression assertions
- Updated aggregate_group_by_tests.rs: GROUP BY COUNT assertions
- Updated aggregate_distinct.rs: DISTINCT COUNT assertions
- Updated aggregate_count_sum_avg_tests.rs: various COUNT assertions
- Updated aggregate_edge_case_tests.rs: edge case assertions

Note: 3 tests still failing due to CASE expression logic issues
with Numeric types (test_count_star_in_simple_case_expression,
test_count_star_in_nested_case_expression, test_count_distinct_vs_count_all).
These require investigation into CASE expression type handling.

Progress: 615 passing (up from 584), 3 failing (down from 34)



* fix: Add numeric type comparison support to CASE expressions (#1182)

* test: Disable flaky and long-running tests to prevent CI hangs

Disables two problematic tests that were causing issues in the test suite:

1. test_select5_no_oom - 10-minute timeout test that's too long for regular CI runs
2. test_timeout_default_is_300 - Race condition with test_timeout_wraps_execution

Both tests are marked with #[ignore] and can still be run explicitly with:
cargo test -- --ignored



* fix: Add numeric type comparison support to CASE expressions

Fixes three failing tests related to CASE expressions with Numeric types:
- test_count_star_in_simple_case_expression
- test_count_star_in_nested_case_expression
- test_count_distinct_vs_count_all

## Root Cause
CASE expressions were comparing Numeric values (returned by COUNT) with
Integer literals, but values_are_equal() didn't support cross-type
numeric comparisons.

## Changes

### Core Fix
- Updated values_are_equal() in crates/executor/src/evaluator/core.rs
  to support numeric type comparisons across all numeric types
  (Integer, Smallint, Bigint, Unsigned, Numeric, Float, Real, Double)
- Uses to_f64() helper to convert both values for comparison
- Handles type coercion properly for CASE expression equality checks

### Test Assertion Updates
Updated test assertions to expect Numeric instead of Integer for
COUNT results:
- crates/executor/src/tests/aggregate_distinct.rs
- crates/executor/tests/having_subquery_tests.rs
- crates/executor/tests/parallel_execution_tests.rs
- crates/executor/tests/predicate_pushdown_tests.rs
- crates/executor/tests/test_predicate_pushdown_integration.rs

All executor tests now pass.



---------


---------

+++ crates/executor/src/select/executor/aggregation/evaluation/aggregate_function.rs
@@ -47,7 +47,7 @@ pub(super) fn evaluate(
                 ));
             }
             // Fast path: COUNT(*) without DISTINCT is just row count (O(1) vs O(n))
-            let result = types::SqlValue::Integer(group_rows.len() as i64);
+            let result = types::SqlValue::Numeric(group_rows.len() as f64);
             // Cache the result
             executor.aggregate_cache.borrow_mut().insert(cache_key, result.clone());
             return Ok(result);
